+++ src/locale/en-US.ts
@@ -1,0 +1,10 @@
+export const locale = {
+  cancel: 'Cancel',
+  confirm: 'Confirm',
+  loading: 'Loading...',
+  submit: 'Submit',
+  pullRefresh: {
+    pulling: 'Pull to refresh...',
+    loosing: 'Loose to refresh...',
+  },
+};