+++ src/locale/zh-CN.ts
@@ -1,0 +1,10 @@
+export const locale = {
+  cancel: '取消',
+  confirm: '确认',
+  loading: '加载中...',
+  submit: '提交',
+  pullRefresh: {
+    pulling: '下拉即可刷新...',
+    loosing: '释放即可刷新...',
+  },
+};