+++ src/tab/index.tsx
@@ -10,8 +10,16 @@ export type TabProps = {
   dot?: boolean;
   info?: number | string;
   disabled?: boolean;
+  lazyRender?: boolean;
+  isActive?: boolean;
 };
 
-export const Tab: preact.FunctionalComponent<TabProps> = () => {
-  return null;
-};
+export class Tab extends preact.Component<TabProps> {
+  private inited = false;
+
+  render(): preact.JSX.Element {
+    const { isActive, lazyRender, children } = this.props;
+    const shouldRender = (this.inited = this.inited || isActive || !lazyRender);
+    return <preact.Fragment>{shouldRender ? children : null}</preact.Fragment>;
+  }
+}