+++ types/closest.d.ts
@@ -1,0 +1,7 @@
+declare module 'closest' {
+  export default function closest(
+    element: Element | EventTarget,
+    selector: string,
+    checkYoSelf?: boolean,
+  ): Element | null;
+}